+++ src/utils/navConfig.js
@@ -9,10 +9,10 @@ const getNavBarType = (pathname) => {
     if (pathname === '/prescription') return 'prescription';
     if (pathname === '/prescription/upload') return 'prescription_upload';
     if (pathname === '/prescription/result') return 'prescription_result';
-    if (pathname === ('/treat-info/result')) return 'precheck';
-    if (pathname ===  '/call-reservation/number') return 'call';
+    if (pathname === '/treat-info/result') return 'precheck';
+    if (pathname ===  '/call-reservation/select') return 'call';
     if (pathname === '/call-reservation/time') return'reservation_time';
-    if (pathname.startsWith ('/call-reservation/result')) return 'reservation_result';
+    if (pathname.startsWith('/call-reservation/result')) return 'reservation_result';
     if (pathname.endsWith('/qr')) return 'qr';
     if (pathname.startsWith('/chat/')) return 'chatroom';
     if (pathname.startsWith('/treat-info-form')) return 'treat-info-form';
